+++ scout/adapter/mongo/case.py
@@ -516,6 +516,12 @@ class CaseHandler(object):
         old_case = self.case_collection.find_one(
             {'_id': case_obj['_id']}
         )
+
+        # if case is updated and was archived or active - make it inactive
+        updated_status = old_case.get('status')
+        if old_case.get('status') in ['archived', 'active']:
+            updated_status = 'inactive'
+
         updated_case = self.case_collection.find_one_and_update(
             {'_id': case_obj['_id']},
             {
@@ -545,7 +551,7 @@ class CaseHandler(object):
                     'research_requested': case_obj.get('research_requested', False),
                     'multiqc': case_obj.get('multiqc'),
                     'mme_submission': case_obj.get('mme_submission'),
-                    'status': old_case.get('status') if  old_case.get('status') != 'archived' else 'inactive'
+                    'status': updated_status
                 }
             },
             return_document=pymongo.ReturnDocument.AFTER