+++ scout/server/blueprints/cases/views.py
@@ -689,7 +689,7 @@ def pin_variant(institute_id, case_name, variant_id):
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
     variant_obj = store.variant(variant_id)
     user_obj = store.user(current_user.email)
-    link = url_for('variants.variant', institute_id=institute_id, case_name=case_name,
+    link = url_for('variant.variant', institute_id=institute_id, case_name=case_name,
                    variant_id=variant_id)
     if request.form['action'] == 'ADD':
         store.pin_variant(institute_obj, case_obj, user_obj, link, variant_obj)
@@ -705,24 +705,33 @@ def mark_validation(institute_id, case_name, variant_id):
     variant_obj = store.variant(variant_id)
     user_obj = store.user(current_user.email)
     validate_type = request.form['type'] or None
-    link = url_for('variants.variant', institute_id=institute_id, case_name=case_name,
+    link = url_for('variant.variant', institute_id=institute_id, case_name=case_name,
                    variant_id=variant_id)
     store.validate(institute_obj, case_obj, user_obj, link, variant_obj, validate_type)
     return redirect(request.referrer or link)
 
 
-@cases_bp.route('/<institute_id>/<case_name>/<variant_id>/causative', methods=['POST'])
-def mark_causative(institute_id, case_name, variant_id):
+@cases_bp.route('/<institute_id>/<case_name>/<variant_id>/<partial_causative>/causative', methods=['POST'])
+def mark_causative(institute_id, case_name, variant_id, partial_causative=False):
     """Mark a variant as confirmed causative."""
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
     variant_obj = store.variant(variant_id)
     user_obj = store.user(current_user.email)
-    link = url_for('variants.variant', institute_id=institute_id, case_name=case_name,
+    link = url_for('variant.variant', institute_id=institute_id, case_name=case_name,
                    variant_id=variant_id)
     if request.form['action'] == 'ADD':
-        store.mark_causative(institute_obj, case_obj, user_obj, link, variant_obj)
+        if 'partial_causative' in request.form:
+            omim_terms = request.form.getlist('omim_select')
+            hpo_terms =  request.form.getlist('hpo_select')
+            store.mark_partial_causative(institute_obj, case_obj, user_obj, link,
+                variant_obj, omim_terms, hpo_terms)
+        else:
+            store.mark_causative(institute_obj, case_obj, user_obj, link, variant_obj)
     elif request.form['action'] == 'DELETE':
-        store.unmark_causative(institute_obj, case_obj, user_obj, link, variant_obj)
+        if eval(partial_causative):
+            store.unmark_partial_causative(institute_obj, case_obj, user_obj, link, variant_obj)
+        else:
+            store.unmark_causative(institute_obj, case_obj, user_obj, link, variant_obj)
 
     # send the user back to the case that was marked as solved
     case_url = url_for('.case', institute_id=institute_id, case_name=case_name)
@@ -781,8 +790,9 @@ def share(institute_id, case_name):
 @cases_bp.route('/<institute_id>/<case_name>/rerun', methods=['POST'])
 def rerun(institute_id, case_name):
     """Request a case to be rerun."""
-    sender = current_app.config['MAIL_USERNAME']
-    recipient = current_app.config['TICKET_SYSTEM_EMAIL']
+    sender = current_app.config.get('MAIL_USERNAME')
+    recipient = current_app.config.get('TICKET_SYSTEM_EMAIL')
+
     controllers.rerun(store, mail, current_user, institute_id, case_name, sender,
                       recipient)
     return redirect(request.referrer)