+++ scout/server/blueprints/variants/controllers.py
@@ -20,17 +20,14 @@ from scout.constants import (
 from scout.constants.acmg import ACMG_CRITERIA
 from scout.constants.variants_export import EXPORT_HEADER, VERIFIED_VARIANTS_HEADER
 from scout.export.variant import export_verified_variants
-from scout.server.utils import institute_and_case, user_institutes
+from scout.server.utils import (institute_and_case, user_institutes, case_append_bam, variant_case)
 from scout.server.links import (add_gene_links, ensembl, add_tx_links)
 from scout.server.blueprints.genes.controllers import gene
 from scout.utils.requests import fetch_refseq_version
 
+from scout.server.blueprints.variant.utils import (predictions)
+
 LOG = logging.getLogger(__name__)
-
-
-class MissingVerificationRecipientError(Exception):
-    pass
-
 
 def variants(store, institute_obj, case_obj, variants_query, page=1, per_page=50):
     """Pre-process list of variants."""
@@ -90,145 +87,9 @@ def str_variants(store, institute_obj, case_obj, variants_query, page=1, per_pag
     # Nothing unique to STRs on this level. Inheritance? yep, you will want it.
 
     # case bam_files for quick access to alignment view.
-    case_append_bam(store, case_obj)
+    case_append_bam(case_obj)
 
     return variants(store, institute_obj, case_obj, variants_query, page, per_page)
-
-def str_variant(store, institute_id, case_name, variant_id):
-    """Pre-process an STR variant entry for detail page.
-
-    Adds information to display variant
-
-    Args:
-        store(scout.adapter.MongoAdapter)
-        institute_id(str)
-        case_name(str)
-        variant_id(str)
-
-    Returns:
-        detailed_information(dict): {
-            'institute': <institute_obj>,
-            'case': <case_obj>,
-            'variant': <variant_obj>,
-            'overlapping_snvs': <overlapping_snvs>,
-            'manual_rank_options': MANUAL_RANK_OPTIONS,
-            'dismiss_variant_options': DISMISS_VARIANT_OPTIONS
-        }
-        """
-
-    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-    variant_obj =  store.variant(variant_id)
-
-    # fill in information for pilup view
-    variant_case(store, case_obj, variant_obj)
-
-    variant_obj['callers'] = callers(variant_obj, category='str')
-
-    # variant_obj['str_ru']
-    # variant_obj['str_repid']
-    # variant_obj['str_ref']
-
-    variant_obj['comments'] = store.events(institute_obj, case=case_obj,
-                                           variant_id=variant_obj['variant_id'], comments=True)
-
-    return {
-        'institute': institute_obj,
-        'case': case_obj,
-        'variant': variant_obj,
-        'overlapping_snvs': overlapping_snvs,
-        'manual_rank_options': MANUAL_RANK_OPTIONS,
-        'dismiss_variant_options': DISMISS_VARIANT_OPTIONS
-    }
-
-def sv_variant(store, institute_id, case_name, variant_id=None, variant_obj=None, add_case=True,
-               get_overlapping=True):
-    """Pre-process an SV variant entry for detail page.
-
-    Adds information to display variant
-
-    Args:
-        store(scout.adapter.MongoAdapter)
-        institute_id(str)
-        case_name(str)
-        variant_id(str)
-        variant_obj(dcit)
-        add_case(bool): If information about case files should be added
-
-    Returns:
-        detailed_information(dict): {
-            'institute': <institute_obj>,
-            'case': <case_obj>,
-            'variant': <variant_obj>,
-            'overlapping_snvs': <overlapping_snvs>,
-            'manual_rank_options': MANUAL_RANK_OPTIONS,
-            'dismiss_variant_options': DISMISS_VARIANT_OPTIONS
-        }
-    """
-    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-
-    if not variant_obj:
-        variant_obj = store.variant(variant_id)
-
-    if add_case:
-        # fill in information for pilup view
-        variant_case(store, case_obj, variant_obj)
-
-    # frequencies
-    variant_obj['frequencies'] = [
-        ('1000G', variant_obj.get('thousand_genomes_frequency')),
-        ('1000G (left)', variant_obj.get('thousand_genomes_frequency_left')),
-        ('1000G (right)', variant_obj.get('thousand_genomes_frequency_right')),
-        ('GnomAD', variant_obj.get('gnomad_frequency')),
-    ]
-
-    if 'clingen_cgh_benign' in variant_obj:
-        variant_obj['frequencies'].append(('ClinGen CGH (benign)', variant_obj['clingen_cgh_benign']))
-    if 'clingen_cgh_pathogenic' in variant_obj:
-        variant_obj['frequencies'].append(('ClinGen CGH (pathogenic)', variant_obj['clingen_cgh_pathogenic']))
-    if 'clingen_ngi' in variant_obj:
-        variant_obj['frequencies'].append(('ClinGen NGI', variant_obj['clingen_ngi']))
-    if 'clingen_mip' in variant_obj:
-        variant_obj['frequencies'].append(('ClinGen MIP', variant_obj['clingen_mip']))
-    if 'swegen' in variant_obj:
-        variant_obj['frequencies'].append(('SweGen', variant_obj['swegen']))
-    if 'decipher' in variant_obj:
-        variant_obj['frequencies'].append(('Decipher', variant_obj['decipher']))
-
-    variant_obj['callers'] = callers(variant_obj, category='sv')
-
-    overlapping_snvs = []
-    if get_overlapping:
-        overlapping_snvs = (parse_variant(store, institute_obj, case_obj, variant) for variant in
-                            store.overlapping(variant_obj))
-
-    # parse_gene function is not called for SVs, but a link to ensembl gene is required
-    for gene_obj in variant_obj.get('genes', []):
-        if gene_obj.get('common'):
-            ensembl_id = gene_obj['common']['ensembl_id']
-            try:
-                build = int(gene_obj['common'].get('build','37'))
-            except Exception:
-                build = 37
-            gene_obj['ensembl_link'] = ensembl(ensembl_id, build=build)
-
-    variant_obj['comments'] = store.events(institute_obj, case=case_obj,
-                                           variant_id=variant_obj['variant_id'], comments=True)
-
-    case_clinvars = store.case_to_clinVars(case_obj.get('display_name'))
-    if variant_id in case_clinvars:
-        variant_obj['clinvar_clinsig'] = case_clinvars.get(variant_id)['clinsig']
-
-    if not 'end_chrom' in variant_obj:
-        variant_obj['end_chrom'] = variant_obj['chromosome']
-
-    return {
-        'institute': institute_obj,
-        'case': case_obj,
-        'variant': variant_obj,
-        'overlapping_snvs': overlapping_snvs,
-        'manual_rank_options': MANUAL_RANK_OPTIONS,
-        'dismiss_variant_options': DISMISS_VARIANT_OPTIONS
-    }
 
 
 def parse_variant(store, institute_obj, case_obj, variant_obj, update=False, genome_build='37',
@@ -285,12 +146,12 @@ def parse_variant(store, institute_obj, case_obj, variant_obj, update=False, gen
                                            variant_id=variant_obj['variant_id'], comments=True)
 
     if variant_genes:
-        variant_obj.update(get_predictions(variant_genes))
+        variant_obj.update(predictions(variant_genes))
         if variant_obj.get('category') == 'cancer':
             variant_obj.update(get_variant_info(variant_genes))
 
     for compound_obj in compounds:
-        compound_obj.update(get_predictions(compound_obj.get('genes', [])))
+        compound_obj.update(predictions(compound_obj.get('genes', [])))
 
     classification = variant_obj.get('acmg_classification')
     if isinstance(classification, int):
@@ -425,7 +286,6 @@ def get_variant_info(genes):
         data['canonical_transcripts'].append(value)
 
     return data
-
 
 def get_predictions(genes):
     """Get sift predictions from genes."""
@@ -1099,15 +959,22 @@ def verification_email_body(case_name, url, display_name, category, subcategory,
 
 def cancer_variants(store, request_args, institute_id, case_name, form):
     """Fetch data related to cancer variants for a case."""
+
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-    variants_query = store.variants(case_obj['_id'], category='cancer', query=form.data).limit(50)
+    per_page = 50
+    skip_count = per_page * max(page - 1, 0)
+    variants_query = store.variants(case_obj['_id'], category='cancer', query=form.data)
+    variant_count = variants_query.count()
+    more_variants = True if variant_count > (skip_count + per_page) else False
+    variant_res = variants_query.skip(skip_count).limit(per_page)
     data = dict(
+        page=page,
+        more_variants=more_variants,
         institute=institute_obj,
         case=case_obj,
         variants=(parse_variant(store, institute_obj, case_obj, variant, update=True) for
-                  variant in variants_query),
+                  variant in variant_res),
         form=form,
-        variant_type=request_args.get('variant_type', 'clinical'),
     )
     return data
 
@@ -1189,44 +1056,6 @@ def get_clinvar_submission(store, institute_id, case_name, variant_id, submissio
         pinned_vars=pinned,
         clinvars = clinvar_submission_objs
     )
-
-
-def variant_acmg(store, institute_id, case_name, variant_id):
-    """Collect data relevant for rendering ACMG classification form."""
-    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-    variant_obj = store.variant(variant_id)
-    return dict(institute=institute_obj, case=case_obj, variant=variant_obj,
-                CRITERIA=ACMG_CRITERIA, ACMG_OPTIONS=ACMG_OPTIONS)
-
-
-def variant_acmg_post(store, institute_id, case_name, variant_id, user_email, criteria):
-    """Calculate an ACMG classification based on a list of criteria."""
-    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-    variant_obj = store.variant(variant_id)
-    user_obj = store.user(user_email)
-    variant_link = url_for('variants.variant', institute_id=institute_id,
-                           case_name=case_name, variant_id=variant_id)
-    classification = store.submit_evaluation(
-        institute_obj=institute_obj,
-        case_obj=case_obj,
-        variant_obj=variant_obj,
-        user_obj=user_obj,
-        link=variant_link,
-        criteria=criteria,
-    )
-    return classification
-
-
-def evaluation(store, evaluation_obj):
-    """Fetch and fill-in evaluation object."""
-    evaluation_obj['institute'] = store.institute(evaluation_obj['institute_id'])
-    evaluation_obj['case'] = store.case(evaluation_obj['case_id'])
-    evaluation_obj['variant'] = store.variant(evaluation_obj['variant_specific'])
-    evaluation_obj['criteria'] = {criterion['term']: criterion for criterion in
-                                  evaluation_obj['criteria']}
-    evaluation_obj['classification'] = ACMG_COMPLETE_MAP.get(evaluation_obj['classification'])
-    return evaluation_obj
-
 
 def upload_panel(store, institute_id, case_name, stream):
     """Parse out HGNC symbols from a stream."""